+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/Hero.cs
@@ -21,10 +21,10 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
             // Archetype
             var HeroArchetype = Helpers.CreateBlueprint<BlueprintArchetype>("HeroArchetype", bp => {
                 bp.LocalizedName = Helpers.CreateString($"HeroArchetype.Name", "Hero");
-                bp.LocalizedDescription = Helpers.CreateString($"HeroArchetype.Description", "Heroes use their newfound powers to for good. After realising the suffering and "
-                    + "despair of the inhabitants of the new world, the hero sets out to bring knowledge from his old world in order to save them.");
-                bp.LocalizedDescriptionShort = Helpers.CreateString($"HeroArchetype.Description", "Heroes use their newfound powers to for good. After realising the suffering and "
-                    + "despair of the inhabitants of the new world, the hero sets out to bring knowledge from his old world in order to save them.");
+                bp.LocalizedDescription = Helpers.CreateString($"HeroArchetype.Description", "Heroes use their newfound powers for good. After realising the suffering and "
+                    + "despair of the inhabitants of the new world, the hero sets out to bring knowledge from their old world in order to save them.");
+                bp.LocalizedDescriptionShort = Helpers.CreateString($"HeroArchetype.Description", "Heroes use their newfound powers for good. After realising the suffering and "
+                    + "despair of the inhabitants of the new world, the hero sets out to bring knowledge from their old world in order to save them.");
                 bp.RemoveSpellbook = false;
                 bp.RemoveFeatures = new LevelEntry[] {
                     Helpers.LevelEntry(1, SneakAttack),