+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -30,6 +30,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist
 
         public static void Add()
         {
+            // TODO: add channel negative energy to villain
             // TODO: add Hero archetype
 
             // TODO: add scaling natural armor, strength and dexterity (like animal companion) to deathsnatcher